requirements.txt
asgiref==3.8.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
dnspython==2.7.0
evdev==1.9.1
greenlet==3.2.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.2
netifaces==0.10.6
pynput==1.8.1
python-engineio==4.12.0
python-socketio==5.13.0
python-xlib==0.33
requests==2.32.3
scapy==2.6.1
setuptools==3.3
simple-websocket==1.1.0
six==1.17.0
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
